services/user_management/requirements.txt
fastapi
uvicorn
pydantic
passlib[argon2,pbkdf2]
pyjwt
python-dotenv
python-multipart
motor